Yolo-Training/Yolo-Training/requirements.txt
ultralytics
fastapi
uvicorn
pillow
python-multipart
aiortc
uvicorn[standard]